pns_web/requirements.txt
arrow>=0.7.0
Django<2.0
Jinja2>=2.8
django-jinja
djangorestframework
haystack